refactor(auth): move auth client to external directory and update imports
Restructure auth client implementation by moving it to the external directory to better reflect its purpose as an external service integration. Update all related imports and test files to reflect this change. Add comprehensive tests for the auth client functionality.

The change improves code organization by separating external service integrations from internal infrastructure code. This makes the codebase more maintainable and follows better architectural practices.
