+++ src/presentation/auth/routes/__tests__/routes.test.ts
@@ -1,12 +1,12 @@
 import { Router } from "express";
 import { AuthRoutes } from "../routes";
 import { AuthDatasource } from "@/infrastructure/datasources/auth.datasource";
-import { AuthClient } from "@/infrastructure/config/auth.client";
+import { AuthClient } from "@/infrastructure/external/auth/auth.client";
 import { AuthController } from "@/presentation/controller/controller";
 
 // Mock dependencies
 jest.mock("@/infrastructure/datasources/auth.datasource");
-jest.mock("@/infrastructure/config/auth.client");
+jest.mock("@/infrastructure/external/auth/auth.client");
 jest.mock("@/presentation/controller/controller");
 jest.mock("@/presentation/middlewares/auth.middleware", () => ({
   AuthMiddleware: {
