+++ src/presentation/auth/routes/routes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { AuthDatasource } from "@/infrastructure/datasources/auth.datasource";
-import { AuthClient } from "@/infrastructure/config/auth.client";
+import { AuthClient } from "@/infrastructure/external/auth/auth.client";
 import { AuthController } from "@/presentation/controller/controller";
 import { AuthMiddleware } from "@/presentation/middlewares/auth.middleware";
 import { CryptoMiddleware } from "@/presentation/middlewares/crypto.middleware";
